リダイレクトの処理をまとめる
