--- app/routes/($lang)._redirects.idea/route.tsx
@@ -1,7 +1,0 @@
-import { redirect } from "@remix-run/cloudflare"
-
-export const loader = async () => {
-  return redirect("/themes", {
-    status: 302,
-  })
-}
