--- app/routes/($lang)._redirects.illust/route.tsx
@@ -1,7 +1,0 @@
-import { redirect } from "@remix-run/cloudflare"
-
-export const loader = async () => {
-  return redirect("/posts/2d", {
-    status: 302,
-  })
-}
