--- app/routes/($lang)._redirects.real/route.tsx
@@ -1,7 +1,0 @@
-import { redirect } from "@remix-run/cloudflare"
-
-export const loader = async () => {
-  return redirect("/posts/3d", {
-    status: 302,
-  })
-}
