--- app/routes/($lang)._redirects.stamp-space/route.tsx
@@ -1,7 +1,0 @@
-import { redirect } from "@remix-run/cloudflare"
-
-export const loader = async () => {
-  return redirect("/stickers", {
-    status: 302,
-  })
-}
